rename metrics to analytics
