+++ analytics/analytics_test.go
@@ -1,4 +1,4 @@
-package metrics
+package analytics
 
 import (
 	"testing"
