+++ analytics/model.go
@@ -1,4 +1,4 @@
-package metrics
+package analytics
 
 type EventType int
